Fixed null removal value from database

+++ MarvBotV3/MarvBotV3/Database/DataAccess.cs
@@ -283,7 +283,7 @@ namespace MarvBotV3.Database
             {
                 var nextRoll =  db.TbNextRoll.AsQueryable().FirstOrDefault(x => x.UserID == 0 || x.UserID == userId);
 
-                if (remove)
+                if (remove && nextRoll != null)
                 {
                     db.TbNextRoll.Remove(nextRoll);
                     db.SaveChanges();
